test(run-protocol): move swingsetsTests group next to vaultFactory group
